piece put to piece area

+++ core/src/com/luzi82/irenepuzzle/Utils.java
@@ -12,11 +12,11 @@ public class Utils {
 
     public static final float PHI = (float)((1+Math.sqrt(5))/2);
 
-    public static void draw(SpriteBatch batch,Texture img,int[] wsen){
+    public static void draw(SpriteBatch batch,Texture img,float[] wsen){
         batch.draw(img,wsen[0],wsen[1],wsen[2]-wsen[0],wsen[3]-wsen[1]);
     }
 
-    public static void draw(SpriteBatch batch, TextureRegion img, int[] wsen){
+    public static void draw(SpriteBatch batch, TextureRegion img, float[] wsen){
         batch.draw(img,wsen[0],wsen[1],wsen[2]-wsen[0],wsen[3]-wsen[1]);
     }
